Add new sub-category: Runtime/Operating Systems
Following discussions in the Special Purpose Operating Systems (SPOS) work group within the Runtime TAG, we've introduced a new “Runtime/Operating Systems” sub-category. This replaces the previous classification, which didn't accurately represent the projects within the group.

The updated sub-category will better align with the SPOS WG projects and accommodate other projects that may join the foundation in the future.

This version is concise and directly conveys the motivation and future scope of the change. Let me know if you’d like to adjust any details!

Signed-off-by: Mauro Morales <contact@mauromorales.com>